+++ src/tiffimage_int.cpp
@@ -1535,12 +1535,11 @@ namespace Exiv2 {
 
     } // TiffCreator::getPath
 
-    ByteOrder TiffParserWorker::decode(
-              ExifData&          exifData,
+    ByteOrder TiffParserWorker::decode(ExifData&          exifData,
               IptcData&          iptcData,
               XmpData&           xmpData,
         const byte*              pData,
-              uint32_t           size,
+              size_t size,
               uint32_t           root,
               FindDecoderFct     findDecoderFct,
               TiffHeaderBase*    pHeader
@@ -1565,10 +1564,9 @@ namespace Exiv2 {
 
     } // TiffParserWorker::decode
 
-    WriteMethod TiffParserWorker::encode(
-              BasicIo&           io,
+    WriteMethod TiffParserWorker::encode(BasicIo&           io,
         const byte*              pData,
-              uint32_t           size,
+              size_t size,
         const ExifData&          exifData,
         const IptcData&          iptcData,
         const XmpData&           xmpData,
@@ -1647,12 +1645,8 @@ namespace Exiv2 {
         return writeMethod;
     } // TiffParserWorker::encode
 
-    TiffComponent::UniquePtr TiffParserWorker::parse(
-        const byte*              pData,
-              uint32_t           size,
-              uint32_t           root,
-              TiffHeaderBase*    pHeader
-    )
+    TiffComponent::UniquePtr TiffParserWorker::parse(const byte* pData, size_t size, uint32_t root,
+                                                     TiffHeaderBase* pHeader)
     {
         if (pData == 0 || size == 0)
             return nullptr;
@@ -1719,7 +1713,7 @@ namespace Exiv2 {
     {
     }
 
-    bool TiffHeaderBase::read(const byte* pData, uint32_t size)
+    bool TiffHeaderBase::read(const byte* pData, size_t size)
     {
         if (!pData || size < 8) return false;
