+++ src/tiffvisitor_int.cpp
@@ -1132,10 +1132,7 @@ namespace Exiv2 {
 
     } // TiffEncoder::add
 
-    TiffReader::TiffReader(const byte*    pData,
-                           uint32_t       size,
-                           TiffComponent* pRoot,
-                           TiffRwState    state)
+    TiffReader::TiffReader(const byte* pData, size_t size, TiffComponent* pRoot, TiffRwState state)
         : pData_(pData),
           size_(size),
           pLast_(pData + size),
